pyproject.toml
[project]
name = "best-shot-extraction"
version = "0.1.0"
description = "Extract best highlight clips from continuous video takes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "open-clip-torch>=2.20.0",
    "opencv-python>=4.8.0",
    "ffmpeg-python>=0.2.0",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "click>=8.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
best-shot = "best_shot_extraction.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipython>=8.0.0",
]

[tool.uv]
index-url = "https://download.pytorch.org/whl/cpu"
extra-index-url = ["https://pypi.org/simple"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.black]
line-length = 88
target-version = ["py310"]